API reference

Public

TrafficAssignment.TrafficAssignment — Module

TrafficAssignment

A Julia package for studying traffic assignment problems, using data from https://github.com/bstabler/TransportationNetworks.

TrafficAssignment.TrafficAssignmentProblem — Type

TrafficAssignmentProblem(instance_name; ...)
TrafficAssignmentProblem(
    instance_name,
    files;
    best_objective,
    toll_factor,
    distance_factor
)

TrafficAssignment.TrafficAssignmentProblem — Type

struct TrafficAssignmentProblem

Fields

instance_name::String

number_of_zones::Int64

number_of_nodes::Int64

first_thru_node::Int64

number_of_links::Int64

init_node::Vector{Int64}

term_node::Vector{Int64}

capacity::Vector{Float64}

link_length::Vector{Float64}

free_flow_time::Vector{Float64}

b::Vector{Float64}

power::Vector{Float64}

speed_limit::Vector{Float64}

toll::Vector{Float64}

link_type::Vector{Int64}

total_od_flow::Float64

travel_demand::Matrix{Float64}

od_pairs::Vector{Tuple{Int64, Int64}}

toll_factor::Float64

distance_factor::Float64

best_objective::Float64

TrafficAssignment.list_instances — Method

list_instances()

Return a list of available instance names.

TrafficAssignment.solve_frank_wolfe — Method

solve_frank_wolfe(td; method, max_iter_no, step, log, tol)

This function implements methods to find traffic equilibrium flows: currently, Frank-Wolfe (FW) method, Conjugate FW (CFW) method, and Bi-conjugate FW (BFW) method.

Settings

method=:fw / :cfw / :bfw (default=:bfw)

step=:exact / :newton: exact line search using golden section / a simple Newton-type step (default=:exact)

log=:on / :off: displays information from each iteration or not (default=:off)

max_iter_no::Integer: maximum number of iterations (default=2000)

tol::Real: tolerance for the Average Excess Cost (AEC) (default=1e-3)

References

Mitradjieva, M., & Lindberg, P. O. (2013). The Stiff Is Moving-Conjugate Direction Frank-Wolfe Methods with Applications to Traffic Assignment. Transportation Science, 47(2), 280-293.

TrafficAssignment.summarize_instances — Method

summarize_instances()

Return a DataFrame summarizing the dimensions of all available instances.

Private

TrafficAssignment.instance_files — Method

instance_files(instance_name)

Return a named tuple (; flow_file, net_file, node_file, trips_file) containing the absolute paths to the 4 data tables of an instance.