API reference

Public

TrafficAssignment.TrafficAssignment — Module

TrafficAssignment

A Julia package for studying traffic assignment problems, using instaces from two datasets:

"TransportationNetworks", available at https://github.com/bstabler/TransportationNetworks

"UnifiedTrafficDataset", available at https://figshare.com/articles/dataset/Aunifiedandvalidatedtrafficdatasetfor20UScities/24235696

TrafficAssignment.TrafficAssignmentProblem — Type

struct TrafficAssignmentProblem{Coord<:Union{Missing, Vector{<:Tuple{var"#s27", var"#s27"} where var"#s27"<:Number}}, Capa<:Number, Length<:Number, Free<:Number, Speed<:Number, BPRMult<:Union{Number, SparseArrays.SparseMatrixCSC{<:Number}}, BPRPow<:Union{Number, SparseArrays.SparseMatrixCSC{<:Number}}, Toll<:Union{Missing, SparseArrays.SparseMatrixCSC{<:Number}}, LinkT<:Union{Missing, SparseArrays.SparseMatrixCSC{<:Integer}}, Flow<:Union{Missing, SparseArrays.SparseMatrixCSC{<:Number}}, Dem<:Number, TF<:Union{Missing, Number}, DF<:Union{Missing, Number}}

Instance of the static traffic assignment problem.

Details

The link travel time is given by the formula of the Bureau of Public Roads (BPR):

t = t0 * (1 + α * (f/c)^β)

where

t is the travel time

t0 is the free flow time

f is the flow along the link

c is the link capacity

α is a multiplicative coefficient (often taken to be 0.15)

β is an exponent (often taken to be 4)

Fields

dataset_name::String: name of the dataset, one of ["TransportationNetworks", "UnifiedTrafficDataset"]

instance_name::String: name of the instance (subfolder inside the dataset)

nb_nodes::Int64: number of nodes in the network (nodes are numbered from 1 to nb_nodes)

nb_links::Int64: number of directed links in the network

real_nodes::UnitRange{Int64}: interval of nodes that correspond to real intersections

zone_nodes::UnitRange{Int64}: interval of nodes that correspond to artificial zones

node_coord::Union{Missing, Vector{<:Tuple{var"#s27", var"#s27"} where var"#s27"<:Number}}: coordinates of the nodes for plotting

valid_longitude_latitude::Bool: whether node_coord corresponds to the longitude and latitude

link_id::SparseArrays.SparseMatrixCSC{Int64, Int64}: matrix of link ids starting at one

link_capacity::SparseArrays.SparseMatrixCSC{Capa, Int64} where Capa<:Number: matrix of link capacities (c in the BPR formula)

link_length::SparseArrays.SparseMatrixCSC{Length, Int64} where Length<:Number: matrix of link lengths

link_free_flow_time::SparseArrays.SparseMatrixCSC{Free, Int64} where Free<:Number: matrix of link free flow times (t0 in the BPR formula)

link_speed_limit::SparseArrays.SparseMatrixCSC{Speed, Int64} where Speed<:Number: matrix of link speed limits

link_bpr_mult::Union{Number, SparseArrays.SparseMatrixCSC{<:Number}}: link multiplicative factors α in the BPR formula, either a single scalar or a matrix

link_bpr_power::Union{Number, SparseArrays.SparseMatrixCSC{<:Number}}: link exponents β in the BPR formula, either a single scalar or a matrix

link_toll::Union{Missing, SparseArrays.SparseMatrixCSC{<:Number}}: matrix of link tolls

link_type::Union{Missing, SparseArrays.SparseMatrixCSC{<:Integer}}: matrix of link types

demand::Dict{Tuple{Int64, Int64}, Dem} where Dem<:Number: demand by OD pair

origins::Vector{Int64}: vector of unique origins for the OD pairs

destinations::Vector{Int64}: vector of unique destinations for the OD pairs

removed_od_pairs::Vector{Tuple{Int64, Int64}}: OD pairs removed because no path between them exists

destination_free_flow_time::Dict{Int64, Vector{Free}} where Free<:Number: dictionary mapping each destination to a vector of free flow path times for every possible origin

toll_factor::Union{Missing, Number}: conversion factor turning toll costs into temporal costs, expressed in time/toll

distance_factor::Union{Missing, Number}: conversion factor turning distance costs into temporal costs, expressed in time/length

optimal_flow::Union{Missing, SparseArrays.SparseMatrixCSC{<:Number}}: provided matrix of optimal link flows

TrafficAssignment.TrafficAssignmentProblem — Method

TrafficAssignmentProblem(dataset_name, instance_name)

User-friendly constructor for TrafficAssignmentProblem.

The dataset must be one of ["TransportationNetworks", "UnifiedTrafficDataset"], the instance can be chosen from list_instances.

Tip

When you run this function for the first time, the DataDeps package will ask you to confirm download. If you want to skip this check, for instance during CI, set the environment variable ENV["DATADEPS_ALWAYS_ACCEPT"] = true.

TrafficAssignment.list_instances — Method

list_instances()
list_instances(dataset_name)

Return a list of the available instances, given as a tuple with their dataset.

TrafficAssignment.nb_links — Method

nb_links(problem::TrafficAssignmentProblem)

Return the number of links in the network (including links to and from zone nodes).

TrafficAssignment.nb_nodes — Method

nb_nodes(problem::TrafficAssignmentProblem)

Return the number of nodes in the network (including zone nodes).

TrafficAssignment.nb_zones — Method

nb_zones(problem::TrafficAssignment)

Return the number of fake nodes in the network that represent zones.

TrafficAssignment.plot_network — Function

plot_network(
    problem::TrafficAssignmentProblem, flow=nothing;
    nodes=false, zones=false, tiles=false
)

Plot a transportation network, possibly on top of real-world map tiles.

If a flow is provided, network edges will be colored according to their congestion level.

Warning

This function requires loading one of Makie.jl's backends beforehand, ideally GLMakie.jl. Using tiles=true requires loading Tyler.jl in addition.

TrafficAssignment.social_cost — Method

social_cost(problem::TrafficAssignmentProblem, flow::AbstractMatrix)

Compute the social cost induced by a matrix of link flows.

TrafficAssignment.solve_frank_wolfe — Method

solve_frank_wolfe(
    problem::TrafficAssignmentProblem,
    frank_wolfe_alg;
    verbose, kwargs...
)

Solve a traffic assignment problem using an algorithm from the FrankWolfe library.

Keyword arguments are passed to frank_wolfe_alg.

TrafficAssignment.summarize_instances — Method

summarize_instances()
summarize_instances(dataset_name)

Return a DataFrame summarizing the dimensions of the available instances inside a datase.

Private

TrafficAssignment.datapath — Method

datapath(dataset_name)
datapath(dataset_name, instance_name)

Return the absolute path to the raw data.

The dataset_name must be one of ["TransportationNetworks", "UnifiedTrafficDataset"].

TrafficAssignment.instance_files — Method

instance_files(dataset_name, instance_name)

Return a named tuple containing the absolute paths to the individual data tables of an instance.

TrafficAssignment.postprocess! — Method

postprocess!(pb)

Perform some data cleaning on a TrafficAssignmentProblem:

precompute free flow times to every destination present in the demand

remove OD pairs without a path between them (the free flow time is infinite)