API reference

Public

TrafficAssignment.TrafficAssignment — Module

TrafficAssignment

A Julia package for studying traffic assignment problems, using data from https://github.com/bstabler/TransportationNetworks.

TrafficAssignment.TrafficAssignmentProblem — Type

TrafficAssignmentProblem(instance_name; ...)
TrafficAssignmentProblem(
    instance_name,
    files;
    toll_factor,
    distance_factor
)

User-friendly constructor for TrafficAssignmentProblem.

The provided instance_name must be one of the subfolders in https://github.com/bstabler/TransportationNetworks.

When you run this function for the first time, the DataDeps package will ask you to confirm download. If you want to skip this check, for instance during CI, set the environment variable ENV["DATADEPS_ALWAYS_ACCEPT"] = true.

TrafficAssignment.TrafficAssignmentProblem — Type

struct TrafficAssignmentProblem{C<:Union{Nothing, Vector{Float64}}, F<:Union{Nothing, SparseArrays.SparseMatrixCSC{Float64, Int64}}}

Instance of the static traffic assignment problem.

Details

The link travel time is given by travel_time = free_flow_time * ( 1 + b * (flow/capacity)^power).

The generalized cost is cost = travel_time + toll_factor * toll + distance_factor * distance.

Fields

instance_name::String

number_of_zones::Int64

number_of_nodes::Int64

first_thru_node::Int64

number_of_links::Int64

init_node::Vector{Int64}

term_node::Vector{Int64}

capacity::SparseArrays.SparseMatrixCSC{Float64, Int64}

link_length::SparseArrays.SparseMatrixCSC{Float64, Int64}

free_flow_time::SparseArrays.SparseMatrixCSC{Float64, Int64}

b::SparseArrays.SparseMatrixCSC{Float64, Int64}

power::SparseArrays.SparseMatrixCSC{Float64, Int64}

speed_limit::SparseArrays.SparseMatrixCSC{Float64, Int64}

toll::SparseArrays.SparseMatrixCSC{Float64, Int64}

link_type::SparseArrays.SparseMatrixCSC{Int64, Int64}

total_od_flow::Float64

travel_demand::Matrix{Float64}

od_pairs::Vector{Tuple{Int64, Int64}}

X::Union{Nothing, Vector{Float64}}

Y::Union{Nothing, Vector{Float64}}

optimal_flow_volume::Union{Nothing, SparseArrays.SparseMatrixCSC{Float64, Int64}}

optimal_flow_cost::Union{Nothing, SparseArrays.SparseMatrixCSC{Float64, Int64}}

toll_factor::Float64

distance_factor::Float64

TrafficAssignment.list_instances — Method

list_instances()

Return a list of available instance names.

TrafficAssignment.plot_network — Method

plot_network(problem)

Plot a transportation network with Makie.

This function requires loading one of Makie's backends beforehand.

TrafficAssignment.plot_network_osm — Method

plot_network_osm(problem)

Plot a transportation network with Makie over OpenStreetMap tiles.

This function requires loading one of Makie's backends beforehand, as well as an internet connection. It assumes the node coordonates are given in degrees.

TrafficAssignment.social_cost — Method

social_cost(problem, flow)

Compute the social cost induced by a matrix of link flows.

TrafficAssignment.solve_frank_wolfe — Method

solve_frank_wolfe(problem; ...)
solve_frank_wolfe(
    problem,
    frank_wolfe_alg;
    verbose,
    kwargs...
)

Solve a traffic assignment problem using an algorithm from the FrankWolfe library.

Keyword arguments are passed to frank_wolfe_alg.

TrafficAssignment.summarize_instances — Method

summarize_instances()

Return a DataFrame summarizing the dimensions of all available instances.

Private

TrafficAssignment.instance_files — Method

instance_files(instance_name)

Return a named tuple (; flow_file, net_file, node_file, trips_file) containing the absolute paths to the 4 data tables of an instance.